requirements.txt
requests
pytest
pyyaml
coloredlogs
mock
python-slugify
Flask
flask_restx
healthcheck
uwsgi
pylint